folders

+++ BetterSelf/Folder/FolderView.swift
@@ -10,19 +10,14 @@ import SwiftUI
 
 struct FolderView: View {
     @Environment(\.modelContext) var modelContext
-    @Query(sort: \Folder.date) var folders: [Folder]
+    @Query(filter: #Predicate<Folder> { $0.isChecked == true},
+           sort: \Folder.date) var folders: [Folder]
 
     @State private var newFolder: Folder?
     @State private var addFolder = false
     @State private var searchText = ""
 
-    var filteredFolders: [Folder] {
-        if searchText.isEmpty {
-           folders
-        } else {
-            folders.filter { $0.name.localizedStandardContains(searchText) }
-        }
-    }
+    @State private var selectedReminder: Reminder?
 
     var body: some View {
         NavigationStack {
@@ -33,36 +28,11 @@ struct FolderView: View {
 
                 Color.purpleOverlayGradient
                     .ignoresSafeArea()
-                List{
 
-                    NavigationLink{
-                        HomeView()
-                    } label: {
-                        HStack {
-                            FolderRowView(folder: nil)
-
-                        }
-                    }
-                    .buttonStyle(.plain)
-                    .listRowBackground(Color.clear)
-                    .listRowInsets(EdgeInsets(top: 8, leading: 16, bottom: 8, trailing: 16))
-
-                    ForEach(filteredFolders){ folder in
-                        NavigationLink{
-                            HomeView(folder: folder)
-                        } label: {
-                            FolderRowView(folder: folder)
-
-                        }
-                    }
-                    .onDelete(perform: deleteFolder)
-                    .listRowBackground(Color.clear)
-                    .listRowInsets(EdgeInsets(top: 8, leading: 16, bottom: 8, trailing: 16))
+                FoldersList(searchText: $searchText, selectedReminder: $selectedReminder)
+                    .searchable(text: $searchText, placement: .navigationBarDrawer,  prompt: "Search")
 
 
-                }
-                .searchable(text: $searchText, placement: .navigationBarDrawer, prompt: "Search")
-                .listStyle(PlainListStyle())
 
             }
             .navigationTitle("BetterSelf")
@@ -83,7 +53,7 @@ struct FolderView: View {
                 }
 
             }
-            
+
             .sheet(isPresented: $addFolder, onDismiss: deleteEmptyFolder){
                 if let folder = newFolder {
                     AddFolderView(folder: folder)
@@ -91,19 +61,23 @@ struct FolderView: View {
                 }
             }
             .toolbarBackground(Color.purpleOverlayGradient, for: .bottomBar, .navigationBar, .tabBar)
+            .navigationDestination(item: $selectedReminder) { reminder in
+                ReminderView(reminder: reminder)
+            }
+
         }
     }
-    func deleteFolder(at offsets: IndexSet) {
-        for offset in offsets {
-            let folder = folders[offset]
-            modelContext.delete(folder)
-        }
-    }
+
     func deleteEmptyFolder() {
         if let folder = newFolder {
+            guard folder.isChecked == false else { return }
             let nameIsNotValid = folders.contains { $0.name.lowercased() == folder.name.lowercased() }
             if folder.name.isEmpty || nameIsNotValid {
                 modelContext.delete(folder)
+
+            }
+            else {
+                folder.isChecked = true
             }
         }
 
@@ -117,4 +91,119 @@ struct FolderView: View {
     FolderView()
 }
 
+
+struct FoldersList: View {
+    @Environment(\.modelContext) var modelContext
+    @Environment(\.isSearching) var isSearching
+    @Query(filter: #Predicate<Folder> { $0.isChecked == true},
+           sort: \Folder.date) var folders: [Folder]
+    @Binding var searchText: String
+
+
+    @Query(filter: #Predicate<Reminder> { $0.isChecked == true
+    }, sort: \Reminder.date) var reminders: [Reminder]
+
+    @Binding var selectedReminder: Reminder?
+
+    var filteredReminders: [Reminder] {
+        if searchText.isEmpty {
+            reminders
+        } else {
+            reminders.filter { $0.title.localizedStandardContains(searchText) }
+        }
+    }
+
+
+
+    var body: some View {
+
+        Group{
+
+            if isSearching || !searchText.isEmpty {
+                List{
+                    ForEach(filteredReminders){ reminder in
+                        Button {
+                            //                                if reminder.type == .InstantInsight && reminder.firebaseVideoURL == nil {
+                            //                                    refuseLoading.toggle()
+                            //                                }
+                            //                                else {
+                            selectedReminder = reminder
+                            //                                }
+                        } label: {
+                            ReminderRowView(reminder: reminder, isPreview: true)
+
+                        }
+                        .listRowBackground(Color.clear)
+                        .listRowInsets(EdgeInsets(top: 4, leading: 16, bottom: 4, trailing: 16))
+                        .listRowSeparator(.hidden)
+                        .buttonStyle(PlainButtonStyle())
+                    }
+                }
+                .listStyle(PlainListStyle())
+            }
+            else {
+
+
+#warning("add a Pinned Area where you can choose Max 3 reminders to be pinned and access them immediately")
+
+                VStack(alignment: .leading, spacing: 10){
+                        Text("Folders")
+                            .font(.title3)
+                            .bold()
+                            .multilineTextAlignment(.leading)
+
+
+                }
+
+                Divider()
+                List{
+
+
+                    NavigationLink{
+                        HomeView()
+                    } label: {
+                        HStack {
+                            FolderRowView(folder: nil)
+
+                        }
+                    }
+                    .buttonStyle(.plain)
+                    .listRowBackground(Color.clear)
+                    .listRowInsets(EdgeInsets(top: 8, leading: 16, bottom: 8, trailing: 16))
+
+                    ForEach(folders){ folder in
+                        NavigationLink{
+                            HomeView(folder: folder)
+                        } label: {
+                            FolderRowView(folder: folder)
+
+                        }
+                    }
+                    .onDelete(perform: deleteFolder)
+                    .listRowBackground(Color.clear)
+                    .listRowInsets(EdgeInsets(top: 8, leading: 16, bottom: 8, trailing: 16))
+
+
+
+
+                }
+                .listStyle(PlainListStyle())
+            }
+
+        }
+
+
+
+
+
+    }
+
+    func deleteFolder(at offsets: IndexSet) {
+        for offset in offsets {
+            let folder = folders[offset]
+            modelContext.delete(folder)
+        }
+    }
+
+}
